+++ .gitignore
@@ -1,1 +1,2 @@
 jsconfig.json
+TODO.txt
