+++ index.html
@@ -90,7 +90,7 @@
 
 		<!-- A set of HTML modals, which can be displayed by JavaScript -->
 		<div id="modals">
-			<mdc-modal id="albumModal" header="Add files to album" data-list="true" data-newtype="child" data-newlabel="New Album" onaccept="Album.getById(this.value).addFiles(app.els.filesCont.selection);"></mdc-modal>
+			<mdc-modal id="albumModal" header="Add files to album" data-list="true" data-newtype="child" data-newlabel="New Album" onaccept="Album.getById(this.value).addFiles(App.app.els.filesCont.selection);"></mdc-modal>
 
 			<mdc-modal id="tagModal" header="Tag files" data-list="true" data-newtype="item" data-newlabel="New Tag" onaccept="setModalInp('tag', 'tagText');"></mdc-modal>
 
@@ -234,7 +234,7 @@
 				</form>
 			</mdc-modal>
 
-			<mdc-modal id="modal-people-face-set" header="Set/Change Person" data-list="true" data-newtype="group" data-newlabel="New Person" onaccept="let modal = this; app.els.filesCont.selection.forEach(function (id) { Face.getById(id).setPerson($(modal).val()); });"></mdc-modal>
+			<mdc-modal id="modal-people-face-set" header="Set/Change Person" data-list="true" data-newtype="group" data-newlabel="New Person" onaccept="let modal = this; App.app.els.filesCont.selection.forEach(function (id) { Face.getById(id).setPerson($(modal).val()); });"></mdc-modal>
 
 			<mdc-modal id="errorModal" header="An Error has Occurred" onaccept="window.location.reload();">
 				<br />
@@ -332,7 +332,7 @@
 						<option onclick="window.location = 'gallery.html' + window.location.search;" title="Switch display mode" disabled>Switch to Gallery mode</option>
 						<option onclick="window.open('settings.html', '_blank');" title="Change configuration settings" disabled>Settings</option>
 						<option onclick="window.open('feedback.html', '_blank');" title="Give feedback or report a bug" disabled>Feedback</option>
-						<option onclick="app.logOut();" title="Log out of your account">Log Out</option>
+						<option onclick="App.app.logOut();" title="Log out of your account">Log Out</option>
 					</mdc-menu>
 
 					<mdc-menu id="toolBar-menu" class="mdc-menu-anchor">
@@ -430,28 +430,28 @@
 						<div id="addressBar-address-box" class="mdc-chip">
 							<p id="addressBar-address">/</p>
 						</div>
-						<div id="search" onkeydown="if (event.key == 'Enter') app.refreshFilesData(null, null, {'search': $(this).find('#searchinput').val()});">
-							<mdc-text id="searchinput" placeholder="Search" data-icon-after="search" data-icon-before="arrow_back" title="Search the current view for files" onsubmit="app.refreshFilesData(null, null, {'search': $(this).val()});"></mdc-text>
+						<div id="search" onkeydown="if (event.key == 'Enter') App.app.refreshFilesData(null, null, {'search': $(this).find('#searchinput').val()});">
+							<mdc-text id="searchinput" placeholder="Search" data-icon-after="search" data-icon-before="arrow_back" title="Search the current view for files" onsubmit="App.app.refreshFilesData(null, null, {'search': $(this).val()});"></mdc-text>
 						</div>
 					</address-bar>
 
 					<!-- Toolbar, containing functions relating to the management and categorisation of files -->
 					<tool-bar id="toolBar" data-axis="max-height" data-size="200px">
-						<button class="tool-button tool-button-small mdc-button mdc-button--raised mdc-theme--secondary-bg mdc-ripple-surface" onclick="app.els.filesCont.selectAll(true);" title="Select all items on the current page">
+						<button class="tool-button tool-button-small mdc-button mdc-button--raised mdc-theme--secondary-bg mdc-ripple-surface" onclick="App.app.els.filesCont.selectAll(true);" title="Select all items on the current page">
 							<i class="material-icons">select_all</i>
 							<span class="toolbar-text">Select All</span>
 						</button>
 
 						<br />
 
-						<button class="tool-button tool-button-small mdc-button mdc-button--raised mdc-theme--secondary-bg mdc-ripple-surface" onclick="app.els.filesCont.selectAll(false);" title="Clear the current selection">
+						<button class="tool-button tool-button-small mdc-button mdc-button--raised mdc-theme--secondary-bg mdc-ripple-surface" onclick="App.app.els.filesCont.selectAll(false);" title="Clear the current selection">
 							<i class="material-icons">clear</i>
 							<span class="toolbar-text">Select None</span>
 						</button>
 
 						<br />
 
-						<button class="tool-button tool-button-small mdc-button mdc-button--raised mdc-theme--secondary-bg mdc-ripple-surface" onclick="app.els.filesCont.invertSelection();" title="Invert which files are selected">
+						<button class="tool-button tool-button-small mdc-button mdc-button--raised mdc-theme--secondary-bg mdc-ripple-surface" onclick="App.app.els.filesCont.invertSelection();" title="Invert which files are selected">
 							<i class="material-icons">invert_colors</i>
 							<span class="toolbar-text">Invert Selection</span>
 						</button>
@@ -461,7 +461,7 @@
 
 					<!-- Sortbar, containing functions relating to the display of files on the page -->
 					<sort-bar id="sortBar">
-						<mdc-tabs onchange="app.setView(this.getTabName());" data-nochange="true">
+						<mdc-tabs onchange="App.app.setView(this.getTabName());" data-nochange="true">
 							<nav id="sortbar-viewswitcher" class="mdc-tab-bar mdc-tab-bar--icon-tab-bar">
 								<a class="mdc-tab mdc-tab--active" value="files" title="Files View">
 									<i class="material-icons mdc-tab__icon">view_module</i>
@@ -480,7 +480,7 @@
 						</mdc-tabs>
 
 						<div id="thumbScaleDiv">
-							<mdc-slider name="thumbScaler" id="thumbScaler" title="Change Thumbnail Size" oninput="app.els.filesCont.scaleFiles(this.value);" onchange="app.config.set('thumb_scale', Math.round(this.value));" ondblclick="this.value = app.config.defaults.thumb_scale.default.desktop; this.oninput(); this.onchange();"></mdc-slider>
+							<mdc-slider name="thumbScaler" id="thumbScaler" title="Change Thumbnail Size" oninput="App.app.els.filesCont.scaleFiles(this.value);" onchange="App.app.config.set('thumb_scale', Math.round(this.value));" ondblclick="this.value = App.app.config.defaults.thumb_scale.default.desktop; this.oninput(); this.onchange();"></mdc-slider>
 						</div>
 
 						<div id="pageNumbers"></div>
@@ -489,7 +489,7 @@
 						<i id="toolBar-show" class="mdc-icon-toggle material-icons" onclick="$('#toolBar').get(0).toggle();" title="Toggle the toolbar">build</i>
 						<i id="filterBar-show" class="mdc-icon-toggle material-icons" onclick="$('#filterBar').get(0).toggle();" title="Toggle the filtering options">filter_list</i>
 
-						<mdc-select name="filecount" id="filecount" label="Files per page" onchange="app.refreshPage(this.value, null);">
+						<mdc-select name="filecount" id="filecount" label="Files per page" onchange="App.app.refreshPage(this.value, null);">
 							<option value="10">10</option>
 							<option value="25">25</option>
 							<option value="50">50</option>
@@ -498,7 +498,7 @@
 							<option value="inf">Unlimited</option>
 						</mdc-select>
 
-						<mdc-select name="select_mode" id="select-mode" label="Selection mode" onchange="app.config.set('select_mode', this.value);">
+						<mdc-select name="select_mode" id="select-mode" label="Selection mode" onchange="App.app.config.set('select_mode', this.value);">
 							<option value="0" title="Click to select, double-click to open">Standard</option>
 							<option value="1" title="Click to open, no selection">View</option>
 							<option value="2" title="Click to select, no opening">Select</option>
