+++ ts/elements/popups/InfoPopup.js
@@ -2,7 +2,7 @@
 class InfoPopup extends HTMLElement {
 	//Setup popup when hover starts
 	setup (item, event) {
-		if (app.config.platform == "mobile") return;
+		if (App.app.config.platform == "mobile") return;
 
 		this.item = item;
 		Input.xPos = event.clientX;
@@ -57,7 +57,7 @@ class InfoPopup extends HTMLElement {
 
 	//Hide popup
 	hide (div, event) {
-		if (app.config && app.config.platform == "mobile" || $(this).css("transform") == "matrix(0, 0, 0, 0, 0, 0)") return;
+		if (App.app.config && App.app.config.platform == "mobile" || $(this).css("transform") == "matrix(0, 0, 0, 0, 0, 0)") return;
 
 		Input.xPos = event.clientX;
 		Input.yPos = event.clientY;
