+++ ts/elements/regions/FilterBar.ts
@@ -10,7 +10,7 @@ class FilterBar extends ToggleBar {
 
 		this.sortable = Sortable.create($("#filterBar-list").get(0), {animation: 150, handle: ".filterhandle", draggable: ".draggable"});
 
-		let filters = app.getQueryParam("filter").split("/");
+		let filters = App.app.getQueryParam("filter").split("/");
 		for (var i in filters) {
 			let newItem = this.add(true);
 
@@ -29,7 +29,7 @@ class FilterBar extends ToggleBar {
 			newItem.find(".filter-value").val(filters[i].substr(2));
 		}
 
-		$("#filterBar-isf-check").get(0).checked = app.getQueryParam("isf");
+		$("#filterBar-isf-check").get(0).checked = App.app.getQueryParam("isf");
 
 		/* if (allData.page_meta.filters == oldData.page_meta.filters) return;
 
@@ -104,7 +104,7 @@ class FilterBar extends ToggleBar {
 			return $(this).find(".filter-action").val() + $(this).find(".filter-aspect").val() + $(this).find(".filter-value").val();
 		}).toArray();
 		let filterStr = filterArray.join("/");
-		app.refreshFilesData(null, null, {"filter": filterStr, "isf": $("#filterBar-isf-check").get(0).checked});
+		App.app.refreshFilesData(null, null, {"filter": filterStr, "isf": $("#filterBar-isf-check").get(0).checked});
 	}
 }
