+++ ts/main.ts
@@ -1,0 +1,3 @@
+import { App } from "./controllers/App";
+
+App.start(document.getElementById("app"));
