+++ ts/models/all_models.ts
@@ -1,0 +1,5 @@
+export { Album } from "./Album"
+export { Face } from "./Face"
+export { FileObject } from "./FileObject"
+export { GeoTagArea, GeoTag } from "./GeoTag"
+export { PersonGroup, Person } from "./Person"
